# web-console/be/src/api-specifications/api-spec.yml
openapi: "3.0.3"
info:
  title: daasiot-dccs
  version: "1.2.1"

  contact:
    name: API Support
    email: a.farfaglia@sebyone.it

  license:
    name: MPL-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
servers:
  - url: http://localhost:{port}/api
    description: Development server
    variables:
      port:
        default: "3000"
        description: default port for the development server
...
components:
  parameters:
    limit:
      name: limit
      in: query
      description: Numero massimo di oggetti da ritornare
      required: false
      schema:
        type: integer
        minimum: 1
        default: 20
      example: 20
    offset:
      name: offset
      in: query
      description: Offset per la paginazione
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
  responses:
    InternalError:
      description: Errore interno del server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Din:
      type: object
      properties:
        id:
          type: integer
          description: ID del DIN
        sid:
          type: string
          description: identificatore della rete daas (SID)
        din:
          type: string
          description: identificatore univoco del nodo daas all'interno della rete
        p_res:
          type: string
          description: preferenze di rete dell'utente, (riservatezza,
          example: "000"
        skey:
          type: string
          description: SKEY del DIN
          example: "9efbaeb2a94f"
...
    DeviceCreationObject:
      type: object
      properties:
        device_model_id:
          type: integer
          description: ID del modello del device
        din_id:
          type: integer
          description: ID del DIN associato al device
        name:
          type: string
          description: Nome del device
        serial:
          type: string
          description: Seriale del device
        latitude:
          type: number
          description: latitudine
          example: 39.3015
        longitude:
          type: number
          description: longitudine
          example: 16.2536
      required:
        - din_id
        - name
        - serial
    Device:
      allOf:
        - $ref: "#/components/schemas/DeviceCreationObject"
        - type: object
          properties:
            id:
              type: integer
              description: ID del Device
    DeviceFull:
      allOf:
        - $ref: "#/components/schemas/Device"
        - type: object
          properties:
            device_model:
              $ref: "#/components/schemas/DeviceModel"
            din:
              $ref: "#/components/schemas/Din"
    DeviceModelCreationObject:
      type: object
      properties:
        device_group_id:
          type: integer
          description: ID del gruppo del modello
        description:
          type: string
          description: descrizione del modello
        name:
          type: string
          description: nome del modello
    DeviceModel:
      allOf:
        - $ref: "#/components/schemas/DeviceModelCreationObject"
        - type: object
          properties:
            id:
              type: integer
              description: ID del modello del device
              nullable: false
...
    DDO:
      description: Daas Data Object
      type: object
      properties:
        id:
          type: integer
          description: ID del DDO
          nullable: false
          example: 3
        din_id_dst:
          type: integer
          description: ID del DIN che ha ricevuto questo DDO
          nullable: false
          example: 2
        din_id_src:
          type: integer
          description: ID del DIN che ha inviato questo DDO
          nullable: false
          example: 1
        timestamp:
          type: string
          format: date-time
          description: Data e orario di quando è stato inviato il DDO
          nullable: true
          example: 2024-07-21T17:32:28Z
        typeset_id:
          type: integer
          nullable: false
          example: 41
        payload:
          type: string
          format: base
          description: Payload contenuto all'interno del DDO, in codifica Base64
          nullable: true
          example: eyJtZXNzYWdnaW8iOiAiY2lhbyBNb25kbyJ9
        payload_size:
          type: integer
          description: Dimensione del payload decodificato in byte, per come è stato ricevuto
          nullable: true
          example: 32
    PaginationData:
      description: Meta dati sulla paginazione dei dati
      properties:
        count:
          type: integer
          description: numero di oggetti contenuti tra i dati della risposta
          minimum: 0
          example: 20
        offset:
          type: integer
          description: gli oggetti contenuti dei dati della risposta partono da questo offset
          minimum: 40
        limit:
          type: integer
          description: numero massimo di oggetti contenuti tra i dati della risposta
          minimum: 0
          example: 20
        total:
          description: numero totale di oggetti per questa query
          type: integer
          minimum: 0
          example: 67
        has_next:
          type: boolean
          description: flag che indica se ci sono altri oggetti da visualizzare
          example: true
        has_prev:
          type: boolean
          description: flag che indica se ci sono oggetti precedenti da visualizzare
          example: false
    Error:
      type: object
      properties:
        error_name:
          type: string
          description: Nome dell'errore
          example: "TypeError"
        message:
          type: string
          description: Messaggio di errore

#region Paths
paths:
...
  /devices:
    get:
      description: mostra una pagina di tutti i devices o una lista filtrata.
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: q
          in: query
          description: >
            Filtro per nome del device.
          schema:
            type: string
          example: "Device 1"

      tags:
        - devices
      responses:
        "200":
          description: Lista di tutti i devices.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Device"
                  pagination:
                    $ref: "#/components/schemas/PaginationData"
        default:
          $ref: "#/components/responses/InternalError"
    post:
      description: Crea un nuovo device, con associato un device model e un DIN.
      tags:
        - devices
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceCreationObject"
              example:
                device_model_id: 4
                din_id: 1
                name: "Device 1"
                latitude: 39.3015
                longitude: 16.2536
      responses:
        "200":
          description: Device aggiunto con successo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Device model o DIN non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"
  /devices/{deviceId}:
    get:
      description: Ritorna un device specifico.
      tags:
        - devices
      parameters:
        - name: deviceId
          in: path
          description: ID del device
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceFull"
        "404":
          description: Device non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"

    put:
      description: Aggiorna un device specifico e il DIN associato.
      tags:
        - devices
      parameters:
        - name: deviceId
          in: path
          description: ID del device
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DeviceCreationObject"
                - type: object
                  properties:
                    din:
                      $ref: "#/components/schemas/Din"
              example:
                name: "device 2024"
                din: { "sid": "100", "din": "87" }
      responses:
        "200":
          description: Device aggiornato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        "400":
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Device model o DIN non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"

    delete:
      description: Elimina un device specifico.
      tags:
        - devices
      parameters:
        - name: deviceId
          in: path
          description: ID del device
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device eliminato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        "404":
          description: Device non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"
  /devices/{deviceId}/ddos/:
    parameters:
      - name: deviceId
        in: path
        description: ID del device che ha ricevuto o inviato i DDO
        required: true
        schema:
          type: integer
          minimum: 0
          example: 1
    get:
      description: Ritorna i ddo ricevuti o inviati da un device. Di default ritorna i DDO ricevuti.
      tags:
        - devices
        - DDOs
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - name: sent
          in: query
          description: Se true ritorna i DDO inviati, altrimenti i DDO ricevuti
          schema:
            type: boolean
          example: true
      responses:
        "200":
          description: Pagina di DDO e informazioni sulla paginazione
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DDO"
                  pagination:
                    $ref: "#/components/schemas/PaginationData"
        "404":
          description: Device non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/InternalError"
  /devices/{deviceId}/reports/:
    parameters:
      - name: deviceId
        in: path
        description: ID del device che ha ricevuto o inviato i DDO
        required: true
        schema:
          type: integer
          minimum: 0
          example: 1
    get:
      description: Ritorna i ddo ricevuti o inviati da un device sotto formato di pdf o excel.
      tags:
        - devices
        - DDOs
      parameters:
        - name: extension
          required: false
          in: query
          description: definisce l'estensione del file, se pdf o xlsx
          schema:
            type: string
            enum: [pdf, xlsx]
            example: pdf
      responses:
        "200":
          description: Pagina di DDO e informazioni sulla paginazione
          content:
            application/pdf:
              schema:
                type: string
                format: binary
              # examples:
              #   sampleFile:
              #     summary: Un esempio di file PDF
              #     externalValue: "../../static/public/img/reporting_templates/esempio_report.pdf"
        "404":
          description: Device non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: Estensione non valida o non supportata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: problema nella generazione del report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  #region Device Models
...